Add files via upload
This is a summary report on the paper ‘ Think Globally, Act Locally: A Deep Neural Network Approach to High-Dimensional Time Series Forecasting’ which was authored by Rajat Sen, Hsiang-Fu Yu, and Inderjit Dhillon while working in Amazon at the time. It was published on October 29, 2019 in the NIPS’19: Proceedings of the 33rd International Conference on Neural Information Processing Systems ,arXiv:1905.03806v2 [stat.ML]. It currently has 163 citations in Google Scholar.